test: refactor/restore 'TestForceMetricsSynchronization'
