+++ test/metrics/metrics_test.go
@@ -2,6 +2,7 @@ package e2e
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"testing"
@@ -20,6 +21,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -644,4 +646,75 @@ func banUser(t *testing.T, hostAwait *wait.HostAwaitility, email string) *toolch
 	err := hostAwait.CreateWithCleanup(t, bannedUser)
 	require.NoError(t, err)
 	return bannedUser
+}
+
+func TestForceMetricsSynchronization(t *testing.T) {
+
+	// given
+	awaitilities := WaitForDeployments(t)
+	hostAwait := awaitilities.Host()
+	hostAwait.UpdateToolchainConfig(t,
+		testconfig.AutomaticApproval().Enabled(true),
+		testconfig.Metrics().ForceSynchronization(false))
+
+	userSignups := CreateMultipleSignups(t, awaitilities, nil, 2)
+
+	// delete the current toolchainstatus/toolchain-status resource and restart the host-operator pod,
+	// so we can start with accurate counters/metrics and not get flaky because of previous tests,
+	// in particular w.r.t the `userSignupsPerActivationAndDomain` counter which is not decremented when a user
+	// is deleted
+	err := hostAwait.DeleteToolchainStatus(t, "toolchain-status")
+	require.NoError(t, err)
+	// restarting the pod after the `toolchain-status` resource was deleted will trigger a recount based on resources
+	err = hostAwait.DeletePods(client.InNamespace(hostAwait.Namespace), client.MatchingLabels{"control-plane": "controller-manager"})
+	require.NoError(t, err)
+	hostAwait.InitMetrics(t, awaitilities.Member1().ClusterName, awaitilities.Member2().ClusterName)
+
+	t.Run("tampering activation-counter annotations", func(t *testing.T) {
+		// change the `toolchain.dev.openshift.com/activation-counter` annotation value
+		for _, userSignup := range userSignups {
+			// given
+			annotations := userSignup.Annotations
+			annotations[toolchainv1alpha1.UserSignupActivationCounterAnnotationKey] = "10"
+			// when
+			mergePatch, err := json.Marshal(map[string]interface{}{
+				"metadata": map[string]interface{}{
+					"annotations": annotations,
+				},
+			})
+			require.NoError(t, err)
+			err = hostAwait.Client.Patch(context.TODO(), userSignup, client.RawPatch(types.MergePatchType, mergePatch))
+			// then
+			require.NoError(t, err)
+		}
+
+		t.Run("verify metrics did not change after restarting pod without forcing recount", func(t *testing.T) {
+			// given
+			hostAwait.UpdateToolchainConfig(t, testconfig.Metrics().ForceSynchronization(false))
+
+			// when restarting the pod
+			err := hostAwait.DeletePods(client.InNamespace(hostAwait.Namespace), client.MatchingLabels{"control-plane": "controller-manager"})
+
+			// then
+			require.NoError(t, err)
+			// metrics have not changed yet
+			hostAwait.WaitForMetricDelta(t, wait.MasterUserRecordsPerDomainMetric, 0, "domain", "external")                       // value was increased by 1
+			hostAwait.WaitForMetricDelta(t, wait.UsersPerActivationsAndDomainMetric, 0, "activations", "1", "domain", "external") // value was increased by 1
+		})
+
+		t.Run("verify metrics are still correct after restarting pod and forcing recount", func(t *testing.T) {
+			// given
+			hostAwait.UpdateToolchainConfig(t, testconfig.Metrics().ForceSynchronization(true))
+
+			// when restarting the pod
+			// TODO: unneeded once the ToolchainConfig controller will be in place ?
+			err := hostAwait.DeletePods(client.InNamespace(hostAwait.Namespace), client.MatchingLabels{"control-plane": "controller-manager"})
+
+			// then
+			require.NoError(t, err)
+			// metrics have been updated
+			hostAwait.WaitForMetricDelta(t, wait.MasterUserRecordsPerDomainMetric, 0, "domain", "external")                        // unchanged
+			hostAwait.WaitForMetricDelta(t, wait.UsersPerActivationsAndDomainMetric, 2, "activations", "10", "domain", "external") // updated
+		})
+	})
 }
